Add variable for file name specification
Now the filename has to start with "WAgroep", and will have whatever string the variable 'filespec' is assigned to added. So to change the text file you want to pars, you can select by changing the variable filespec. That string is also added when saving the WApostclasslist with pickle.dump. So text file and results of parsing can be identified as belonging together.